move test class - CsvFormatterTests
